### 2018-06-12 157671f
Refactor tile to derive output CRS from input CRS by default. Only reproject when provided new crs_out argument. Also add profile argument. Update verison, documentation.
