+++ NEWS.md
@@ -1,3 +1,11 @@
+# tiler 0.2.9.9000
+
+* Refactored `tile` to no longer force reproject projected rasters from `file` to EPSG 4326. Moving forward, output CRS is derived from input CRS by default. To force a reprojection of `file` prior to tiling to any CRS, including EPSG 4326, provide the output Proj4 string to `crs_out`.
+* Another argument was added, `profile`, for the tile-cutting profile used internally by `gdal2tiles`. This defaults to `mercator` but can also be set to `raster` or `geodetic` as needed depending on the output CRS derived from `file`, specified in `crs`, or provided for reprojection via `crs_out`.
+* Updated documentation and readme.
+* Updated unit tests.
+* Updated vignette.
+
 # tiler 0.2.0 (Release date: 2018-06-09)
 
 * All three `gdal2tiles*` scripts have been updated to accept a command line argument when called by R that provides a path for any temporary files, i.e., `tmp.*.vrt` files created by the `gdal2tiles*` scripts. These were previously accumulating in the system temp folder. The new temporary directory is a sub-directory inside `tempdir()`. Therefore, it is cleaned up when exiting R. Nevertheless, `tile` also force deletes this subdirectory immediately after its internal system call to one of the `gdal2tiles*` scripts returns, so the temporary sub-directory does not even exist for the full duration of the `tile` call.
