### 2018-06-12 2cf23b5
Update documentation, unit tests, vignette.
