+++ vignettes/tiler.Rmd
@@ -14,7 +14,13 @@ knitr::opts_chunk$set(
 tmpfiles <- list.files(tempdir(), full.names = TRUE) # any pre-existing temp files
 ```
 
-The `tiler` package provides a map tile-generator function for creating map tile sets for use with packages such as `leaflet`.
+The `tiler` package generates map tiles for:
+
+* standard Web Mercator, e.g. EPSG 4326/EPSG 3857
+* other geospatial map projections
+* non-geospatial, simple coordinate reference systems
+
+`tiler` provides a map tile-generator function for creating map tile sets for use with packages such as `leaflet`.
 
 In addition to generating map tiles based on a common raster layer source, it also handles the non-geographic edge case, producing map tiles from arbitrary images. These map tiles, which have a "simple CRS", a non-geographic simple Cartesian coordinate reference system, can also be used with `leaflet` when applying the simple CRS option. Map tiles can be created from an input file with any of the following extensions: `tif`, `grd` and `nc` for spatial maps and `png`, `jpg` and `bmp` for basic images.
 
@@ -29,6 +35,12 @@ You should not rely on a gratuitous check like this; `tiler` is being overly gen
 ## Geographic map tiles
 
 ### Context
+
+The overarching perspective of `tiler` is isolation of tile creation from tile usage and map creation. Depending on the source map, generating map tiles can be inconveniently slow to do on the fly for some map use cases. It can also result in the creation of tens of thousands of map tiles, or more. A good strategy is to create map tiles, then serve them online at a preferred location. After all of that scaffolding is in place, then separately and subsequently use those map tiles in various applications as an externalized resource.
+
+`tiler` also does not assume what type of maps you will use, e.g., Leaflet, Google Maps, or something else. Therefore, it will not default to producing Web Mercator map tiles. Instead, the output CRS for tiles is derived from the input CRS of the source map file by default, unless some other output CRS is specified by the user.
+
+### Examples
 
 For the sake of simple, expedient examples, the map tiles generated below all use small zoom level ranges. There is also no reason to attempt displaying the tiles here. To make these examples more informative, each raster is loaded and plotted for context, though this is not necessary to the tiling process. Loading the `raster` package is only needed here for the print and plot calls.
 
@@ -63,14 +75,14 @@ Note that these examples rendered to HTML here do not capture the parts of the l
 
 ### Projected maps
 
-The previous example used a map with a geospatial coordinate reference system (CRS) but it was not projected. That map would be ready to view with the `leaflet` package for example, as would the tiles generated based on it. The next example uses a similar map that is projected. In order to generate map tiles that can be used with `leaflet` in the standard CRS, the map must be reprojected first. Then the same map tiles are generated as before.
+The previous example used a map with a geospatial coordinate reference system (CRS) but it was not projected. That map would be ready to view with the `leaflet` package for example, as would the tiles generated based on it. The next example uses a similar map that is projected. In order to generate map tiles that can be used with `leaflet` in the standard CRS, EPSG 4326, the map must be reprojected first. Then the same map tiles are generated as before.
 
 ```{r ex2}
 map <- system.file("maps/map_albers.tif", package = "tiler")
 (r <- raster(map))
 plot(r)
 
-tile(map, tile_dir, "0-3")
+tile(map, tile_dir, "0-3", crs_out = "EPSG:4326")
 ```
 
 ```{r unlink2, echo=FALSE}
@@ -78,6 +90,20 @@ unlink(tile_dir, recursive = TRUE, force = TRUE)
 ```
 
[…]
