### 2018-06-12 5e5681f
Update package descriptions.
