### 2018-06-15 46ce452
Do not force read raster into memory. Force unlink.
