### 2018-06-15 5ae2610
Add metadata helpers.
