+++ NAMESPACE
@@ -1,6 +1,10 @@
 # Generated by roxygen2: do not edit by hand
 
 export(tile)
+export(tile_bounds)
+export(tile_meta)
+export(tile_origin)
+export(tile_res)
 export(tile_viewer)
 export(tiler_options)
 export(view_tiles)
