### 2018-06-15 a83a9c2
Skip test on Appveyor.
