+++ tests/testthat/test-meta.R
@@ -2,11 +2,13 @@ context("metadata")
 
 tmpfiles <- list.files(tempdir()) # any pre-existing temp files
 
-x <- system.file("maps/map_wgs84.tif", package = "tiler")
-tiles <- file.path(tempdir(), "tiles")
-tile(x, tiles, "2-3")
-
 test_that("tile metadata returns as expected", {
+  skip_on_appveyor()
+
+  x <- system.file("maps/map_wgs84.tif", package = "tiler")
+  tiles <- file.path(tempdir(), "tiles")
+  tile(x, tiles, "2-3")
+
   meta <- tile_meta(tiles)
   o <- tile_origin(tiles)
   b <- tile_bounds(tiles)
