### 2018-06-15 942ea65
Update function and unit test.
