### 2018-06-15 e182f83
Improve OSGeo4W.bat detection.
