### 2018-06-15 cb17b51
Update code and documentation regarding Windows behavior and availability of OSGeo4W.
