+++ R/tile.R
@@ -120,7 +120,7 @@ tile <- function(file, tiles, zoom, crs = NULL, crs_out = NULL, profile = c("mer
                  normalizePath(tiles), "\"")
   }
   cat("Creating tiles. Please wait...\n")
-  system(ex, ignore.stderr = TRUE)
+  system(ex)
   unlink(g2t_tmp_dir, recursive = TRUE, force = TRUE)
   if(viewer){
     cat("Creating tile viewer...\n")
