+++ .travis.yml
@@ -1,8 +1,13 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
+language: python
+virtualenv:
+  system_site_packages: true
+
 language: R
 sudo: false
 cache: packages
+
 addons:
   apt:
     packages:
