### 2018-06-18 504292d
Update .travis.yml.
