+++ .travis.yml
@@ -3,11 +3,9 @@
 language: python
 virtualenv:
   system_site_packages: true
-
 language: R
 sudo: false
 cache: packages
-
 addons:
   apt:
     packages:
