### 2018-06-18 ropensci/tiler#6 @leonawicz
Python-gdal on Travis and Appveyor
GDAL remains unavailable on Travis and Appveyor. While tests and examples using `tile` can run without creating tiles, tests and examples will fail if they require access to outputs of `tile`, as they will not exist. I have not yet figured out how to make the required libraries and bindings available in Travis and Appveyor via `.travis.yml` and `appveyor.yml`. Help appreciated.
