### 2018-06-19 882cde0
tiles path can be a url for tile_meta.
