### 2018-06-21 #226 @leonawicz
@karthik Thank you and no rush :)

Also, please feel free to ignore the unpublished (to CRAN) `projections` branch of the repo. Unless you think that is something assistance could be provided for. If you do consider the `projections` branch, it has unresolved issues, but its documentation is kept up to date with its code development. Otherwise, the master branch is what has actually been published.
