### 2018-06-26 #226 @karthik
### Editor checks:

---

#### Editor comments

👋 @leonawicz The package looks to be in really good shape and no substantial issues emerged from the checks. The only minor suggestion is to break up long lines of code to improve readability, but we can wait to see what concrete suggestions the reviewers have.

No worries on the `projections` branch. Reviewers typically only address the master unless you say otherwise.

```
── GP tiler ────────────────────────────────────────────────────────────────────

It is good practice to

    R/tile.R:75:1
    R/tile.R:79:1
    R/tile.R:83:1
    R/tile.R:96:1
    R/tile.R:97:1
    ... and 22 more lines
```

You can go ahead and add a rOpenSci review badge to your readme. It will autoupdate as the review progresses.

---

Reviewers: Currently seeking
Due date: TBD
