### 2018-06-26 512c1b6
Add ROpenSci badge.
