### 2018-06-26 #226 @leonawicz
@karthik Thanks! Badge added.

I can shorten the longer lines (I think my default lintr check allows 120 instead of 80). Thanks.

Yeah, please ignore the `projections` branch then. It is worth taking a peak at to get a sense of future scope, but I have no date set for when I would have the issues resolved and a new version out.
