### 2018-07-16 #226 @jasdumas
Estimated hours spent reviewing: **2.5 Hours**

- Start: 9:50 AM , End: 12:15 PM
---

#### Manual review of GitHub Repo, vignette & installation directions
- The use of package down for the docs is great!
- It would be great to see a clearer **A statement of need** in the README. Maybe just incorporating your comments from the [review issue would be helpful](https://github.com/ropensci/onboarding/issues/226#issue-331620236)
  - Does this package fill a gap that leaflet does not provide (static vs. dynamic tiles)?
  - Who is the intended user base?
  - Also with such a visual package, it would be nice to include a image in the README of one of the resultant maps/tiles from the vignette as a teaser for GitHub.
- CRAN & Development versions downloads successfully without error/messages
- It would be nice to see a seperate section of linux/mac installation directions even though the last sentence of the introduction vignette states, as the only call out to Windows made it appear that this package would only work for windows users:
  > Linux and Mac users should not have to do any additional setup as long as Python and gdal for Python are installed.

- It would be helpful to explicitly state a minimum Python 2.x or Python 3.x versions in the readme/introduction vignette, but I documented my efforts below to get the dependency gdal:
   - It also seems that gdal works with python 2 using `pip install GDAL`
   - My default `python` is `Python 3.4.4 |Anaconda custom (x86_64)| (default, Jan  9 2016, 17:30:09)[GCC 4.2.1 (Apple Inc. build 5577)] on darwin` and for `python3` is `Python 3.5.2 (v3.5.2:4def2a2901a5, Jun 26 2016, 10:47:25) [GCC 4.2.1 (Apple Inc. build 5666) (dot 3)] on darwin`
   - I updated my system-wide python3 to 3.7 and downloaded python2.7.15 to alleviate some download pains
   - I created a virtual envrionment `python3 -m venv ropensci` & `source ropensci/bin/activate` to start
   - I'm still getting errors for `pip3 install GDAL` or `pip install GDAL`
   - Updated setuptools: `pip install --upgrade setuptools`
   - ok turns out that the first example in the introduction vignette works regardless of all my python mistakes :smile:
- The context section is really helpful for performance considerations!

#### Automated tests
_via opening up the Development :package: R project with packrat in a new RStudio session_

- `devtools::check(document=FALSE)`:

> Packages suggested but not available for checking:
  ‘lintr’ ‘covr’ ‘jpeg’ ‘bmp’
R CMD check results
0 errors | 0 warnings | 1 note

> R CMD check succeeded

- I installed `install.packages(c("lintr", "covr", "jpeg", "bmp"))` and the NOTE went away :tada:

- `devtools::test()`:

> ══ Results ════════════════════════════════════════════════════════

>OK:       29

- `goodpractice::gp()`: For the unit tests Goodpractice notes, i think its fine as you have some helper functions that don't need unit tests given the major functions do have unit tests. the long lines note wa already mentioned from @karthik 's editor check. The `rgdal` callout in the DESCRIPTION file might elude to needing a roxygen comment `#' @import rgdal` in one of your functions if you are leveraging a function with `::`.

```
 ── GP tiler ───────────────────────────────────────────────────────

  ✖ write unit tests for all functions, and all
    package code in general. 88% of code lines are covered
    by test cases.

    R/tile.R:79:NA
    R/tile.R:80:NA
    R/tile.R:83:NA
    R/tile.R:84:NA
    R/tile.R:87:NA

  ✖ avoid long code lines, it is bad for
    readability. Also, many people prefer editor windows
    that are about 80 characters wide. Try make your lines

    R/tile.R:75:1
    R/tile.R:79:1
    R/tile.R:83:1
    R/tile.R:96:1
    R/tile.R:97:1

  ✖ fix this R CMD check NOTE: Namespace in
    Imports field not imported from: ‘rgdal’ All declared
    Imports should be used.
```

- `spelling::spell_check_package()`: These 'mispellings' appear to just be industry nomenclature :tada:

```
DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
  WORD                 FOUND IN
Albers               tiler.Rmd:152
bmp                  tiler.Rd:16
                     description:5
CRS                  tile_viewer.Rd:14,16
                     tile.Rd:17,37,45
                     tiler.Rd:14
                     description:4
                     tiler.Rmd:19,66,82,84,98,144,148,165
EPSG                 tile.Rd:55
gdal                 description:6,7
GDAL                 tiler_options.Rd:26,28,31
georeference         tiler.Rmd:165
georeferencing       tile_viewer.Rd:18
                     tile.Rd:25
                     tiler.Rmd:179
geospatial           tile_viewer.Rd:14,16
                     tile.Rd:67
                     tiler.Rd:9
                     description:1
                     tiler.Rmd:66,84,109,144,148
geotiff              tiler.Rmd:152
grd                  tiler.Rd:16
                     description:5
https                tiler.Rd:19
                     description:7
jpg                  tiler.Rd:16
                     description:5
JS                   tile.Rd:73
                     tiler.Rmd:165
knitr                tiler.Rmd:2
nc                   tiler.Rd:16
                     description:5
osgeo                tiler.Rd:19
                     description:7
OSGeo                tiler_options.Rd:26,28
                     description:7
                     tiler.Rmd:23,23
png                  tiler.Rd:16
                     description:5
proj                 tile.Rd:39
Proj                 tile.Rd:17
QGIS                 tiler_options.Rd:28
                     tiler.Rmd:23
rasterized           tile.Rd:57
rasters              tile.Rd:62
                     tiler.Rmd:126,128
reprojected          tile.Rd:55
                     tiler.Rmd:66
reprojection         tile.Rd:27,55,56,57,58,62
                     tiler.Rmd:98,109
reprojects           tiler.Rmd:80
RGBA                 tile.Rd:58,60,61
                     tiler.Rmd:126
rmarkdown            tiler.Rmd:2
tif                  tiler.Rd:16
                     description:5
TMS                  tile.Rd:19,65,67
                     tiler.Rmd:175
trac                 tiler.Rd:19
                     description:7
unprojected          tile.Rd:61
VignetteEncoding     tiler.Rmd:2
VignetteEngine       tiler.Rmd:2
VignetteIndexEntry   tiler.Rmd:2
XYZ                  tile.Rd:19,65,68
                     tiler.Rmd:175
```

- For extra fun, I also ran the [`gramr` package](https://github.com/ropenscilabs/gramr) on the pkgdown RMarkdown file, `write_good_file("/Users/jasminedumas/Desktop/R-directory/ropensci_package_reviews/packrat/src/tiler/leonawicz-tiler-512c1b6/vignettes/tiler.Rmd")`. Full disclosure, I developed this package at the rOpenSci Unconf17, so take the suggestions from this check with a _strong grain of salt_, but other than that the site is great and well written! :smile:

```
| index| offset|reason                                                    |
|-----:|------:|:---------------------------------------------------------|
|   530|     11|"In addition" is wordy or unneeded                        |
|   809|      7|"be used" may be passive voice                            |
|   883|     10|"be created" may be passive voice                         |
|  1067|     11|"is required" may be passive voice                        |
|  1164|      6|"simply" can weaken meaning                               |
|  1242|      5|"it is" is wordy or unneeded                              |
|  1785|      6|"is set" may be passive voice                             |
|  1815|      5|"it is" is wordy or unneeded                              |
|  1916|      4|"only" can weaken meaning                                 |
|  2121|      5|"It is" is wordy or unneeded                              |
|  2124|     14|"is recommended" may be passive voice                     |
|  2148|      6|"simply" can weaken meaning                               |
|  2372|      4|"just" can weaken meaning                                 |
|  2433|      9|"be needed" may be passive voice                          |
|  2490|     10|"additional" is wordy or unneeded                         |
|  2547|     13|"are installed" may be passive voice                      |
|  2705|      8|"There is" is unnecessary verbiage                        |
|  2820|      9|"is loaded" may be passive voice                          |
|  2939|      4|"only" can weaken meaning                                 |
|  3124|      4|"very" is a weasel word and can weaken meaning            |
|  3135|     11|"in order to" is wordy or unneeded                        |
|  3147|      8|"minimize" is wordy or unneeded                           |
|  3193|      7|"quickly" can weaken meaning                              |
|  3263|     10|"be applied" may be passive voice                         |
|  3395|      4|"very" is a weasel word and can weaken meaning            |
|  3474|     11|"a number of" is wordy or unneeded                        |
|  3615|      5|"it is" is wordy or unneeded                              |
|  3618|     14|"is recommended" may be passive voice                     |
|  3661|      4|"only" can weaken meaning                                 |
|  3848|     13|"are generated" may be passive voice                      |
|  3997|      4|"only" can weaken meaning                                 |
|  4135|     10|"subsequent" is wordy or unneeded                         |
|  4171|      8|"are used" may be passive voice                           |
|  4596|     10|"subsequent" is wordy or unneeded                         |
|  4622|      5|"it is" is wordy or unneeded                              |
|  4991|      6|"it was" is wordy or unneeded                             |
|  5168|     12|"is projected" may be passive voice                       |
|  5182|     11|"In order to" is wordy or unneeded                        |
|  5222|      7|"be used" may be passive voice                            |
|  5279|     14|"be reprojected" may be passive voice                     |
|  5325|     13|"are generated" may be passive voice                      |
|  5698|      7|"be seen" may be passive voice                            |
|  6299|      6|"likely" can weaken meaning                               |
|  6311|      9|"is wanted" may be passive voice                          |
|  6370|      5|"it is" is wordy or unneeded                              |
|  6401|     11|"was dropped" may be passive voice                        |
|  6486|     13|"are generated" may be passive voice                      |
|  7041|      6|"be set" may be passive voice                             |
|  7141|      9|"is needed" may be passive voice                          |
|  7185|      4|"just" can weaken meaning                                 |
|  7337|      6|"easily" can weaken meaning                               |
|  7344|      7|"be done" may be passive voice                            |
|  7522|     10|"are passed" may be passive voice                         |
|  7646|      7|"usually" can weaken meaning                              |
|  7654|     10|"be ignored" may be passive voice                         |
|  8187|     13|"are supported" may be passive voice                      |
|  8377|     11|"are colored" may be passive voice                        |
|  8448|      8|"prior to" is wordy or unneeded                           |
|  8523|      6|"simply" can weaken meaning                               |
|  8608|     11|"are ignored" may be passive voice                        |
|  8631|      7|"type of" is wordy or unneeded                            |
|  9159|      9|"is geared" may be passive voice                          |
|  9240|      7|"However" is wordy or unneeded                            |
|  9296|     12|"are required" may be passive voice                       |
|  9409|      7|"usually" can weaken meaning                              |
|  9609|      8|"There is" is unnecessary verbiage                        |
|  9727|      8|"are said" may be passive voice                           |
|  9813|      7|"however" is wordy or unneeded                            |
| 10051|      9|"was shown" may be passive voice                          |
| 10061|     10|"previously" can weaken meaning and is wordy or unneeded  |
| 10139|     13|"was processed" may be passive voice                      |
| 10245|      5|"It is" is wordy or unneeded                              |
| 10253|     10|"previously" can weaken meaning and is wordy or unneeded  |
| 10411|      6|"all of" is wordy or unneeded                             |
| 10428|      8|"be tiled" may be passive voice                           |
| 10711|      8|"There is" is unnecessary verbiage                        |
| 10881|      7|"be used" may be passive voice                            |
| 10965|      8|"is based" may be passive voice                           |
| 11393|     14|"is recommended" may be passive voice                     |
| 11683|      5|"It is" is wordy or unneeded                              |
| 11686|     10|"is assumed" may be passive voice                         |
| 11918|     10|"Additional" is wordy or unneeded                         |
| 12137|      4|"only" can weaken meaning                                 |
| 12400|      6|"simply" can weaken meaning                               |
| 12559|     11|"adjacent to" is wordy or unneeded                        |
| 12819|      4|"only" can weaken meaning                                 |
| 12924|      7|"Finally" can weaken meaning                              |
| 12959|     10|"additional" is wordy or unneeded                         |
| 13053|      9|"be served" may be passive voice                          |
| 13310|     11|"exclusively" can weaken meaning and is wordy or unneeded |
| 13350|      4|"just" can weaken meaning                                 |
| 13848|      8|"are used" may be passive voice                           |
| 14022|     10|"be created" may be passive voice                         |
| 14164|      6|"easily" can weaken meaning                               |
| 14171|      9|"be loaded" may be passive voice                          |
| 14340|      7|"be done" may be passive voice                            |
| 14384|      4|"just" can weaken meaning                                 |
| 14575|      7|"However" is wordy or unneeded                            |
| 14609|     10|"is created" may be passive voice                         |
| 14628|      9|"be viewed" may be passive voice                          |
| 14788|     11|"in order to" is wordy or unneeded                        |
| 15138|     15|"trial and error" is a cliche                             |
```
