### 2018-07-17 a7d60d0
Update readme, docs.
