### 2018-07-17 #226 @leonawicz
Thank you so much @jasdumas ! This is really helpful. I've included a `Motivation` section in the readme regarding statement of need now. I've also added an example screenshot of a tiled map.

`gramr` looks super cool! I can definitely use that in general. I'll try to take a closer look when I have some time.

Re: installation on different systems. I don't know how to do it for Mac and don't have access to any Macs. For Linux, the only systems I have access to are server environments that already have everything Python and geospatial installed that could possibly be needed (and I'm not the sys admin so I don't know those details).

I do use Linux with Travis CI but in that environment it's not truly testing for the successful creation of map tile files. E.g., `tile` can execute and not make any files as a result of system requirements not present and this is not an error. I do not know how to do this yet: [Python-gdal on Travis and Appveyor](https://github.com/leonawicz/tiler/issues/6)

I'll take a closer look at some of the lines not covered by unit tests. Some lines can be difficult to test.

Thanks!
