### 2018-08-18 #226 @karthik
Thanks @leonawicz, I'll take a look at those issues.

Also @Paula-Moraga is going to be our second reviewer and she will complete her review shortly.
