### 2018-08-18 #226 @Paula-Moraga
---

The `tiler` package enables to build map tiles from geospatial and image files. The tiles created can be used with other packages such as `leaflet`. Some examples of tiles produced using Star Trek galaxy maps and used with `leaflet` are [here](https://leonawicz.github.io/rtrek/articles/sc.html) and these are awesome!

The package works fine and the documentation is clear. The package requires Python and the `gdal` library for Python. This can be tricky to install for some users depending on the OS. When using Windows I needed to add some environment variables and I think these steps need to be better explained in the documentation. Also, it would be great to include in the documentation the code to create a tile and actually use it to produce a leaflet map. Other minor issues and results of the automated tests are below.

Congrats on the package!

I use Windows 10 and I needed to download and install `OSGeo4W64`.

To make the package work I needed to execute this:
`tiler_options(python = "C:/OSGeo4W64/bin/python.exe", osgeo4w = "C:/OSGeo4W64/OSGeo4W.bat")`

Then I tried the package adding the Windows environment variables instead. As it is specified in the documentation, I added C:\OSGeo4W64 to the Windows environment variable PATH so R can find OSGeo4W.bat. Then I executed `tile(map, tile_dir, "0-3")` and I got a Warning message with status 127. I fixed that by adding C:\OSGeo4W64\bin to the Windows environment variable PATH so R can find python.exe.

Then I executed `tile(map, tile_dir, "0-3")` again and I got a Warning message with status 1. I fixed that by creating the environment variable PYTHONHOME with value C:\OSGeo4W64\apps\Python27/ That way R can find the local python installation path.

- I would like to see an example of the code needed to create a tile and actually use it in a leaflet map.

- In section Local preview I would change `project/tiles` by `tile_dir` as in the previous examples.

- I create tiles with `tile()` and view them with `view_tiles()`. I do not understand what `tile_viewer()` is for.

- I would write the name of functions with parentheses at the end. For example, `tile()` instead of `tile`.

- The documentation shows some simple examples, and says more difficult tiles can take a long time to run. I suggest you add a table with the time it takes to run different examples and the computer specifications.

#### Automated tests

`devtools::check()`

```
R CMD check results
0 errors | 0 warnings | 1 note
checking package dependencies ... NOTE
Package suggested but not available for checking: 'bmp'
```

`devtools::test()`

```
Loading tiler
Loading required package: testthat
Testing tiler
√ | OK F W S | Context
x |  0 1 1   | lints
test-lintr.R:4: warning: Package Style
incomplete final line found on 'C:/Users/moragasp/tiler-master/tiler-master/.lintr'

test-lintr.R:4: error: Package Style
Invalid DCF format.
Regular lines must have a tag.
Offending lines start with:
  inst/.lintr
1: lintr::expect_lint_free() at C:\Users\moragasp\tiler-master\tiler-master/tests/testthat/test-lintr.R:4
2: lint_package(...)
3: read_settings(path)
4: read.dcf(config_file, all = TRUE)
5: stop(gettextf("Invalid DCF format.\nRegular lines must have a tag.\nOffending lines start with:\n%s",
       paste0("  ", lines, collapse = "\n")), domain = NA)

test-tile.R:2: warning: (unknown)
package ‘raster’ was built under R version 3.4.4

test-tile.R:2: warning: (unknown)
package ‘sp’ was built under R version 3.4.4
√ |  5       | viewer

== Results =====================================================================

OK:       28
```

`goodpractice::gp()`

```

    R/tile.R:79:NA
    R/tile.R:80:NA
    R/tile.R:83:NA
    R/tile.R:84:NA
    R/tile.R:87:NA

1: In readLines(file) :
  incomplete final line found on 'C:/Users/moragasp/tiler-master/tiler-master/.lintr'
2: In MYPREPS[[prep]](state, quiet = quiet) : Prep step for linter failed.

```

`spelling::spell_check_package()`

```
DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
```

`spelling::spell_check_files("README.Rmd")`

```
  WORD             FOUND IN
AppVeyor         README.Rmd:25
codecov          README.Rmd:26
CRS              README.Rmd:34
georeferencing   README.Rmd:45
github           README.Rmd:2
gitter           README.Rmd:28
notfying         README.Rmd:64
OSGeo            README.Rmd:69
Rdoc             README.Rmd:23
```
