### 2018-08-23 #226 @leonawicz
@Paula-Moraga thanks for your review! :) Sorry for delays, work has been very busy lately.

The function `tile_viewer` is for edge cases where the `preview.html` file associated with a set of map tiles needs to be created because it is missing/deleted or because it must be re-created. For example, if someone runs `tile` multiple times with different zoom level settings and `resume = TRUE`, the `preview.html` is created every time `tile` is run. But the final time it is created when running `title`, it will not include all zoom levels. Running `tile_viewer` afterward with the full zoom range specified will re-create the file again, with proper full zoom level range. But no need to create tiles which already exist.

I will look into adding a visible `leaflet` example that demonstrates the resulting tiles soon. This is ideal, I just haven't had time to do yet. Also fixing the `tile_dir` specification in the Local preview section of the documentation.

I'm currently at my limit for the moment in terms of system installation knowledge/experience. I only have one Windows 10 system to test installation and use. It worked for me but I do not have the broader OS experience to be able to determine what an ideal generic solution for Windows installation that works for all users would be. I am even more limited with my knowledge about how to install system requirements on Linux at the moment, and Mac is not even in the picture for me. This is a problematic situation, but unfortunately not one I have been able to resolve on my own yet.
