### 2018-08-25 #226 @Paula-Moraga
@leonawicz Thanks for the `tile_viewer()` explanation. Maybe it is useful to add this explanation in the vignette. I agree it is hard to write installation guidelines that work for everybody. Looking forward to the leaflet example!
