### 2018-08-29 4c619ce
Add leaflet examples, update documentation.
