### 2018-08-29 e467ab9
Increment to dev version, add leaflet package to Suggests, due to new Leaflet examples being added to vignette.
