+++ NEWS.md
@@ -1,3 +1,7 @@
+# tiler 0.2.0.9000
+
+* Added `leaflet` examples of remotely hosted tiles generated by `tiler` to vignette.
+
 # tiler 0.2.0 (Release date: 2018-06-09)
 
 * All three `gdal2tiles*` scripts have been updated to accept a command line argument when called by R that provides a path for any temporary files, i.e., `tmp.*.vrt` files created by the `gdal2tiles*` scripts. These were previously accumulating in the system temp folder. The new temporary directory is a sub-directory inside `tempdir()`. Therefore, it is cleaned up when exiting R. Nevertheless, `tile` also force deletes this subdirectory immediately after its internal system call to one of the `gdal2tiles*` scripts returns, so the temporary sub-directory does not even exist for the full duration of the `tile` call.
