+++ DESCRIPTION
@@ -34,4 +34,4 @@ Imports:
     raster,
     png
 VignetteBuilder: knitr
-RoxygenNote: 6.0.1
+RoxygenNote: 6.1.0
