### 2018-08-29 ca627f9
Update leaflet example.
