+++ vignettes/tiler.Rmd
@@ -200,9 +200,9 @@ Geographic provider tiles based on the low resolution example map included in `t
 library(leaflet)
 tiles <- "https://leonawicz.github.io/tiles/us48lr/tiles/{z}/{x}/{y}.png"
 leaflet(
-  options = leafletOptions(minZoom = 3, maxZoom = 7), 
-    width = "100%", height = "400px") %>% 
-  addTiles(tiles)
+  options = leafletOptions(minZoom = 0, maxZoom = 7), width = "100%", height = "400px") %>% 
+  addProviderTiles("Stamen.Toner") %>% 
+  addTiles(tiles, options = tileOptions(opacity = 0.8)) %>% setView(-100, 40, 3)
 ```
 
 #### Non-geographic
