### 2018-08-29 2e15271
Update Leaflet examples.
