### 2018-08-29 838a0e4
Update Leaflet examples.
