### 2018-08-29 #226 @leonawicz
@Paula-Moraga I still have to update the vignette with a mention about `tile_viewer`, but for now please see the updated vignette regarding the fun Leaflet examples. :)

I have added two examples, one geographic and one non-geographic, using `leaflet` to display remotely hosted tiles that were originally created with `tiler`. The code is shown for the original calls to `tile` that generated these tiles as well as the `leaflet` code for using them in the interactive maps shown.

These are located in the final section "Serving map tiles," in a Leaflet Examples subsection right before the Local Preview subsection. The Local Preview section is where I will also mention `tile_viewer` when I get to that next. Let me know what you think. Thanks!
