### 2018-09-02 #226 @Paula-Moraga
@leonawicz The leaflet examples are not working for me. Coult it be that the variables
`tiles <- "https://leonawicz.github.io/tiles/us48lr/tiles/{z}/{x}/{y}.png"` and `tiles <- "https://leonawicz.github.io/tiles/st2/tiles/{z}/{x}/{y}.png"` are not right?
