### 2018-09-04 #226 @leonawicz
I have been checking here after using pkgdown and hosting online with the GitHub repo: https://leonawicz.github.io/tiler/articles/tiler.html

The Leaflet widgets in the `tiler.html` file are also displaying for me locally.

Are you seeing something different?
