### 2018-09-04 #226 @Paula-Moraga
Yes, that was the reason! I do not see them in RStudio's preview pane but when I click the right panel it works perfectly! The examples you chose are great, thank you!
