### 2018-11-01 3882ce2
Update installation text.
