### 2018-11-01 a2c3a65
Fix bug that was causing OSGeo4W directory on Windows system path to be ignored. Update documentation, simplifying the Windows setup details.
