### 2018-11-01 508fec7
Increment package version, update news.
