+++ NEWS.md
@@ -1,8 +1,10 @@
-# tiler 0.2.0.9000
+# tiler 0.2.1
 
+* Improved and simplified instructions and expectations for Windows use. Windows users must add `OSGeo4W.bat` path to system path.
 * Added `leaflet` examples of remotely hosted tiles generated by `tiler` to vignette.
+* Bug fix related to system path to `OSGeo4W.bat` being ignored on Windows.
 
-# tiler 0.2.0 (Release date: 2018-06-09)
+# tiler 0.2.0
 
 * All three `gdal2tiles*` scripts have been updated to accept a command line argument when called by R that provides a path for any temporary files, i.e., `tmp.*.vrt` files created by the `gdal2tiles*` scripts. These were previously accumulating in the system temp folder. The new temporary directory is a sub-directory inside `tempdir()`. Therefore, it is cleaned up when exiting R. Nevertheless, `tile` also force deletes this subdirectory immediately after its internal system call to one of the `gdal2tiles*` scripts returns, so the temporary sub-directory does not even exist for the full duration of the `tile` call.
 * Added functions `tile_viewer` and `view_tiles` and other supporting functions for generating HTML Leaflet tile preview web page.
@@ -10,7 +12,7 @@
 * Added unit tests.
 * Updated vignette.
 
-# tiler 0.1.6 (Release date: 2018-06-06)
+# tiler 0.1.6
 
 * Made minor formatting changes per CRAN request for resubmission.
 
