+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: tiler
-Version: 0.2.0.9000
+Version: 0.2.1
 Title: Create Geographic and Non-Geographic Map Tiles
 Description: Creates geographic map tiles from geospatial map files or non-geographic map tiles from simple image files. 
     This package provides a tile generator function for creating map tile sets for use with packages such as 'leaflet'. 
@@ -8,9 +8,7 @@ Description: Creates geographic map tiles from geospatial map files or non-geogr
     Map tiles can be created from an input file with any of the following extensions: tif, grd and nc for spatial maps and png, jpg and bmp for basic images.
     This package requires 'Python' and the 'gdal' library for 'Python'. 
     'Windows' users are recommended to install 'OSGeo4W' (<https://trac.osgeo.org/osgeo4w/>) as an easy way to obtain the required 'gdal' support for 'Python'.
-Authors@R: c(person("Matthew", "Leonawicz", email = "mfleonawicz@alaska.edu", role = c("aut", "cre")),
-    person("Scenarios Network for Alaska and Arctic Planning", role = c("cph", "fnd"))
-    )
+Authors@R: person("Matthew", "Leonawicz", email = "matt_leonawicz@esource.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0001-9452-2771"))
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
