### 2018-11-01 f8abb2f
Updtae CRAN comments for next release.
