### 2018-11-05 #226 @karthik
@leonawicz 👋 I just wanted to check in on the status of this review.  Have all of the issues raised by @Paula-Moraga been addressed now?
