### 2018-11-05 #226 @leonawicz
@karthik @Paula-Moraga I think so??? I've done everything I planned on doing

There are some other issues I mentioned earlier that I am unable to do. See https://github.com/ropensci/onboarding/issues/226#issuecomment-414060516 for example of something I really can't figure out, but don't know how critical it is- I defer to your judgement on the necessity. https://github.com/leonawicz/tiler/issues/6

Side note: I pushed a small Windows-only bug fix to CRAN last week which is pending (v0.2.1).
