### 2018-11-05 #226 @karthik
Thanks @leonawicz! I see that all issues have now been addressed. I'll proceed with accepting your package.

Congrats @leonawicz, your submission has been approved! 🎉 Thank you for submitting and @jasdumas and @Paula-Moraga  for thorough and timely reviews. To-dos:

```
[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)

```
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. (We'll turn on the services on our end as needed)

Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/technotes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). If you are, @stefaniebutland will be in touch about content and timing.
