### 2018-11-06 2c292dd
Repo transfer: Update urls, logo, docs to point to ROpenSci.
