+++ pkgdown/_pkgdown.yml
@@ -1,4 +1,4 @@
-url: https://leonawicz.github.io/
+url: https://ropensci.github.io/
 
 template:
   params:
