+++ DESCRIPTION
@@ -13,8 +13,8 @@ License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
 ByteCompile: true
-URL: https://github.com/leonawicz/tiler
-BugReports: https://github.com/leonawicz/tiler/issues
+URL: https://github.com/ropensci/tiler
+BugReports: https://github.com/ropensci/tiler/issues
 SystemRequirements: Python (>= 2.7), python-gdal library (For Windows, gdal installed via OSGeo4W <https://trac.osgeo.org/osgeo4w/> recommended)
   clipboard
 Suggests:
