### 2018-11-08 fc72d06
Fix logo.
