### 2018-11-16 #226 @stefaniebutland
Hello @leonawicz and congratulations. [This link](https://ropensci.org/tags/onboarding/) will give you many examples of blog posts by authors of onboarded packages. In case you are considering writing one, here are some technical and editorial guidelines: https://github.com/ropensci/roweb2#contributing-a-blog-post.

No obligation to do this. Happy to answer any questions.
